+++ src/Permission.psm1
@@ -282,3 +282,5 @@ Export-ModuleMember -Function @('Add-CachedCimInstance','Add-CacheItem','Add-Per
 
 
 
+
+